genai-sven-zimmer/pyproject.toml
[project]
name = "genai-sven-zimmer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "gradio>=5.34.2",
    "openai>=1.91.0",
    "openai-agents>=0.1.0",
    "pandas>=2.3.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "tabulate>=0.9.0",
]
